requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
librosa
numpy
soundfile
yt-dlp
requests
openai-whisper
